python/m_inc/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "m_inc"
version = "0.1.1"
description = "Mercenaries Incorporated - Economic layer for CuBFF"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Apache-2.0"}
authors = [
    {name = "CuBFF M|inc Team"}
]
keywords = ["artificial-life", "economics", "simulation", "bff", "emergence"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Life",
]

dependencies = [
    "pyyaml>=6.0",
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "hypothesis>=6.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
minc = "m_inc.cli:main"

[project.urls]
Homepage = "https://github.com/paradigms-of-intelligence/cubff"
Documentation = "https://github.com/paradigms-of-intelligence/cubff/tree/main/docs/0.1.1"
Repository = "https://github.com/paradigms-of-intelligence/cubff"

[tool.setuptools.packages.find]
where = ["."]
include = ["m_inc*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=m_inc --cov-report=term-missing"
